requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
cmake==3.29.5.1
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
filelock==3.15.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
grpcio==1.64.1
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lit==18.1.7
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==24.1
pandas==2.2.2
-e git+ssh://github.com:X1AOX1A/SeTAR.git@main#egg=peft&subdirectory=src/peft
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
-e git+ssh://github.com:X1AOX1A/SeTAR.git@main#egg=svd_ood&subdirectory=src/svd_ood
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==1.0.3
tokenizers==0.15.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
transformers==4.37.2
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
